Fix strange duplication behavior
It seems that when an entity is parented and spawned in SetMass is called on the entity setting it's mass to 50,000 but never back down to whatever it's supposed to be... Yet somehow it *does* become the weight it's supposed to be, but not via SetMass.
This causes the change in weight on duplicated parented entities to appear to be 50,000-OriginalWeight (A negative number).

Added sanity check in SetMass to ignore changes in mass that result in a negative total mass to circumvent strange behavior.