Fixed elastics breaking cframe during undo
Elastics demonstrate some strange behavior when undone by behaving as if one of the entities it is constrained to is removed before the constraint is. This is not the case with any other constraint (all constraints have prop:DeleteOnRemove(self) called on their respective props). Issue resolved by allowing CFrame to handle props being removed before the constraint instead of just the expected behavior of constraints always being removed before props.